SPI Flash dump from CU300 WiFi module
Scrubbed dump of SPI flash from CU300 WiFi module. MCU is a Marvell 88MW300, external flash is 32Mbit. String analysis shows 2 instances of the Marvell firmware with the 4 byte header of MRVL... With the limited information about this MCU, I am under the impression that there should be 3 segments. It seems the first segment should start at 0x240, but it is obfuscated with some sequence of mirroring a special word length.