Output from CU300 WiFi module P2 pin 6.
Scrubbed of my personal information. This is a different unit than what the SPI flash was captured from. This cap was taken with the main board removed from the robot, and running on the bench. Some information that was scrubbed in the flash dump has been left intact in this cap.