Logic Analyzer capture from CU300 P2, Sigrock Pulseview
D0 - D5 is P2 pins 1 - 6 respectivly. 
D0 is Power, D1 is Ground, D2 is Rx from STM32, D3 is Tx to STM32, D4 is ... I don't know, D5 is Tx to terminal.
I would assume D4 to be Rx from the terminal, but there is no response from keyboard input. The initial pulse-train that is output right after power on is interesting, and reminds me of IR encoding... I will need to look into it.
If I can get the CU300 to respond to keyboard input... that would be great!