Dump From Ecovac Deebot N79
These are the registers I could dump from an Ecovac Deebot N79 board I picked up. RDP1 is enabled on these too, but they did not assign the SWDIO pin to a different function.
Dumped with a Segger J-link, and I have been trying to run the firmware extractor race condition exploit on it, but running into some issues.